refactor: Comment 엔티티 내의 Id 컬럼 타입 Long 으로 변경

+++ src/main/java/net/devgrr/interp/ia/api/comment/entity/Comment.java
@@ -22,7 +22,7 @@ public class Comment extends BaseEntity {
     private Long id;
 
     @Schema(description = "부모 댓글 ID")
-    private Integer parentCommentId;
+    private Long parentCommentId;
 
     @Schema(description = "댓글 내용")
     @Column(nullable = false)
